# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1",
        "title": "MyMovies Application",
        "description": "Information about movies in my collection."
    },
...
    "schemes": [
        "http"
    ],
    "host": "localhost:8080",
    "basePath": "/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "api_key": {
            "description": "A an API-key.",
            "type": "apiKey",
            "name": "api_key",
            "in": "query"
        }
    },
    "definitions": {
...
        "indices": {
            "description": "A list of indices.",
            "type": "array",
            "items": {
                "$ref": "#/definitions/index"
            }
        },
...
        "episodes": {
            "description": "A list of episodes.",
            "type": "array",
            "items": {
                "$ref": "#/definitions/episode"
            }
        },
...
        "title": {
            "type": "object",
            "description": "A movie title.",
            "required": [
                "imdbid",
                "title",
                "year",
                "imageUrl"
            ],
            "properties": {
                "imdbid": {
                    "description": "The imdbid of the movie.",
                    "type": "string"
                },
                "title": {
                    "description": "The movie title.",
                    "type": "string"
                },
                "year": {
                    "description": "The movie release-year.",
                    "type": "string"
                },
                "imageUrl": {
                    "description": "The url o\tf the movie poster-image.",
                    "type": "string"
                }
            },
            "additionalProperties": false
        },
        "titles": {
            "description": "A list of titles.",
            "type": "array",
            "items": {
                "$ref": "#/definitions/title"
            }
        },
...
        "cast": {
            "description": "A list of the movie-cast",
            "type": "array",
            "items": {
                "type": "string"
            }
        },
...
        "keys": {
            "description": "A list of keys.",
            "type": "array",
            "items": {
                "type": "string"
            }
        }
    },
    "paths": {
        "/title/{imdbId}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_TITLES"
                ],
                "summary": "Individual movie title.",
                "description": "Request title of a single movie.",
                "parameters": [
                    {
                        "name": "imdbId",
                        "description": "imdbId. Request a title with imdbId.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the title.",
                        "schema": {
                            "$ref": "#/definitions/title"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/movie/{imdbId}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_DETAILS"
                ],
                "summary": "Individual movie.",
                "description": "Request details on a single movie.",
                "parameters": [
                    {
                        "name": "imdbId",
                        "description": "imdbId. Request a movie with imdbId.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the movie.",
                        "schema": {
                            "$ref": "#/definitions/movie"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "MOVIES",
                    "MOVIE_DETAILS"
                ],
                "summary": "Individual movie.",
                "description": "Delete a single movie.",
                "parameters": [
                    {
                        "name": "imdbId",
                        "description": "imdbId. Delete a movie with imdbId.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the movie."
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/cast/{imdbId}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_DETAILS"
                ],
                "summary": "Cast of a movie.",
                "description": "Request a list of the cast of a single movie.",
                "parameters": [
                    {
                        "name": "imdbId",
                        "description": "imdbId. Request the cast with imdbId.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the cast.",
                        "schema": {
                            "$ref": "#/definitions/cast"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/titles/{index}/{key}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_TITLES"
                ],
                "summary": "Titles associated with an index.",
                "description": "Request a list of titles associated with an index.",
                "parameters": [
                    {
                        "name": "index",
                        "description": "index. The name of the index.",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "all",
                            "actor",
                            "director",
                            "genre",
                            "rating",
                            "tag",
                            "title",
                            "year"
                        ]
                    },
                    {
                        "name": "key",
                        "description": "key. Request a list of titles associated with this key. Must relate to index.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the titles.",
                        "schema": {
                            "$ref": "#/definitions/titles"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/index/keys/{name}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_INDICES"
                ],
                "summary": "Index keys associated with an index-name.",
                "description": "Request a list of index-keys associated with an index-name.",
                "parameters": [
                    {
                        "name": "name",
                        "description": "name. Request a list of index-keys associated with a name.",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "all",
                            "actor",
                            "director",
                            "genre",
                            "rating",
                            "tag",
                            "title",
                            "year"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the keys.",
                        "schema": {
                            "$ref": "#/definitions/keys"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/index/{name}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_INDICES"
                ],
                "summary": "Indices associated with an index-name.",
                "description": "Request a list of indices associated with a name.",
                "parameters": [
                    {
                        "name": "name",
                        "description": "name. Request a list of indices associated with a name.",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "actor",
                            "director",
                            "genre",
                            "rating",
                            "tag",
                            "title",
                            "year"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the indices.",
                        "schema": {
                            "$ref": "#/definitions/indices"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/index/{name}/{key}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_INDICES"
                ],
                "summary": "Indices associated with an index-name and key.",
                "description": "Request a list of indices associated with a name and key.",
                "parameters": [
                    {
                        "name": "name",
                        "description": "name. Request an index associated with a name.",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "actor",
                            "director",
                            "genre",
                            "rating",
                            "tag",
                            "title",
                            "year"
                        ]
                    },
                    {
                        "name": "key",
                        "description": "The index key. Must relate to name.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the index.",
                        "schema": {
                            "$ref": "#/definitions/index"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/episodes/{imdbid}/{season}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_EPISODES",
                    "MOVIE_SEASONS"
                ],
                "summary": "Episodes associated with a show and season.",
                "description": "Request a list of episodes associated with a show and season.",
                "parameters": [
                    {
                        "name": "imdbid",
                        "description": "imdbid. Request a list of episodes associated with imdbid.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "season",
                        "description": "season. Filter the list of episodes by season.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the episodes.",
                        "schema": {
                            "$ref": "#/definitions/episodes"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "MOVIES",
                    "MOVIE_EPISODES",
                    "MOVIE_SEASONS"
                ],
                "summary": "One ore more episodes.",
                "description": "Delete all episodes for single season of a movie.",
                "parameters": [
                    {
                        "name": "imdbId",
                        "description": "imdbId. Delete all episodes for a season with imdbId.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the episodes."
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/episodes/{imdbid}/{season}/{episodeNumber}": {
            "delete": {
                "tags": [
                    "MOVIES",
                    "MOVIE_EPISODES",
                    "MOVIE_SEASONS"
                ],
                "summary": "One episode.",
                "description": "Delete one episodes for single season of a movie.",
                "parameters": [
                    {
                        "name": "imdbId",
                        "description": "imdbId. Delete one episode for a season with imdbId.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the episode."
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/episodes/{imdbid}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_EPISODES"
                ],
                "summary": "All episodes associated with a show.",
                "description": "Request a list of episodes associated with a show.",
                "parameters": [
                    {
                        "name": "imdbid",
                        "description": "imdbid. Request a list of episodes associated with imdbid.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the episodes.",
                        "schema": {
                            "$ref": "#/definitions/episodes"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "MOVIES",
                    "MOVIE_EPISODES"
                ],
                "summary": "One ore more episodes.",
                "description": "Delete all episodes for a movie.",
                "parameters": [
                    {
                        "name": "imdbId",
                        "description": "imdbId. Delete all episodes for a movie with imdbId.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted the episodes."
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/seasons/{imdbid}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_SEASONS"
                ],
                "summary": "A list of seasons associated with a show.",
                "description": "Request a list of seasons associated with a show.",
                "parameters": [
                    {
                        "name": "imdbid",
                        "description": "imdbid. Request a list of seasons associated with imdbid.",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the seasons.",
                        "schema": {
                            "$ref": "#/definitions/seasons"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/movies/{index}": {
            "get": {
                "tags": [
                    "MOVIES",
                    "MOVIE_DETAILS"
                ],
                "summary": "Movies associated with an index.",
                "description": "Request a list of movies associated with an index.",
                "parameters": [
                    {
                        "name": "index",
                        "description": "index. The name of the index.",
                        "in": "path",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "actor",
                            "director",
                            "genre",
                            "rating",
                            "tag",
                            "title",
                            "year"
                        ]
                    },
                    {
                        "name": "name",
                        "description": "name. Request a list of movies associated with this name. Must relate to index.",
                        "in": "query",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the movies.",
                        "schema": {
                            "$ref": "#/definitions/movies"
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    }
}